fix: use sandboxed paths for Mac App Store build
- Create PlatformPaths.GetPolyPilotDir() that redirects to
  FileSystem.AppDataDirectory on MACCATALYST and iOS/Android,
  and to home directory (~/.polypilot/) on desktop builds.
- Update CopilotService.GetPolyPilotBaseDir() to use PlatformPaths
- Update ConnectionSettings.GetPolyPilotDir() to use PlatformPaths
- Remove both /.polypilot/ and /.copilot/ temporary-exception
  entitlements from Entitlements.AppStore.plist

This allows the Mac App Store build to work within the sandbox
without requiring broad home-directory access. The sandbox
automatically remaps HOME to ~/Library/Containers/<bundle-id>/Data/,
so UserProfile-based paths in the bundled copilot CLI will resolve
inside the container.
