+++ global.json
@@ -1,0 +1,5 @@
+{
+  "sdk": {
+    "version": "10.0.102"
+  }
+}
